requirements.txt
streamlit
psycopg2-binary
redis
python-dotenv
Pillow
ollama
